Refactor: Scales and evaluations count display & related courses
